refactor(website): show the original-data hint above the upload section
Move the note out of the small print in the left-hand column of the
"Sequences and metadata" section and render it as its own paragraph
directly under the input mode tabs, where it is actually read.

Keeping it in DataUploadForm means groupId no longer has to be threaded
through FormOrUploadWrapper and SequenceEntryUpload, so those files (and
the FormOrUploadWrapper spec) go back to their original state.
