+++ website/src/components/Submission/FormOrUploadWrapper.tsx
@@ -43,7 +43,6 @@ type FormOrUploadWrapperProps = {
     setFileFactory: Dispatch<SetStateAction<FileFactory | undefined>>;
     setSubmissionFileMapping: Dispatch<SetStateAction<Result<SubmissionFileMapping, Error> | undefined>>;
     organism: string;
-    groupId: number;
     action: UploadAction;
     metadataTemplateFields: Map<string, InputField[]>;
     submissionDataTypes: SubmissionDataTypes;
@@ -62,7 +61,6 @@ export const FormOrUploadWrapper: FC<FormOrUploadWrapperProps> = ({
     setFileFactory,
     setSubmissionFileMapping,
     organism,
-    groupId,
     action,
     metadataTemplateFields,
     submissionDataTypes,
@@ -168,7 +166,6 @@ export const FormOrUploadWrapper: FC<FormOrUploadWrapperProps> = ({
         return (
             <SequenceEntryUpload
                 organism={organism}
-                groupId={groupId}
                 action={action}
                 metadataFile={metadataFile}
                 setMetadataFile={setMetadataFile}
